refactor - parsing was separate into own module.
